feat: Yazi File explorer
Usage: `y` launches yazi and you navigate with vim keybinds
While in yazi pressing `z` is attached to zoxide to quickly
navigate. Pressing `a` (add) while in a directory opens
a section to name the file or directory. There's much more
check it out!

This has been a game changer for my workflow